Slight restructuring / naming
