+++ Helper/initializeProject().swift
@@ -1,5 +1,5 @@
 //
-//  Initialize.swift
+//  initializeProject().swift
 //  Helper
 //
 //  Created by James Bean on 12/26/16.
@@ -8,11 +8,16 @@
 
 import Foundation
 
-func initialize() {
-
-    // createFileStructure()
+func initializeProject(name: String, in directory: URL) throws {
+    try createFileStructure(projectName: name, in: directory)
     // configureProject()
     
+}
+
+func createFileStructure(projectName: String, in directory: URL) throws {
+    // create folders for main and test target
+    try createDirectory(name: projectName, in: directory)
+    try createDirectory(name: "\(projectName).tests", in: Project.directory)
 }
 
 func configureProject() {
